sdk: consolidate silo utils
