+++ projects/sdk/src/lib/silo/Withdraw.ts
@@ -3,7 +3,7 @@ import { Token } from "src/classes/Token";
 import { TokenValue } from "src/classes/TokenValue";
 import { BeanstalkSDK } from "../BeanstalkSDK";
 import { DepositCrate, Silo } from "../silo";
-import { sortCratesBySeason } from "../silo.utils";
+import { sortCratesBySeason } from "./utils";
 import { pickCrates } from "./utils";
 
 export class Withdraw {
